refactor cell count cube
